refactor: CarList 클래스 분리
